+++ app/MMB/README.md
@@ -1,1 +1,9 @@
-## Write me
+This package implement Manage My Booking (MMB) feature.
+
+```
+src/
+├── components/                       - generic components reused accross all scenes in MMB
+├── navigation/                       - React Navigation related files
+├── scenes/                           - navigation independent (more or less) scenes
+└── MainMenu.js                       - main file with defined items
+```
