+++ app/MMB/index.js
@@ -1,8 +1,26 @@
 // @flow
 
+import * as React from 'react';
 import { AppRegistry } from 'react-native';
+import { AuthContext } from '@kiwicom/mobile-relay';
+import { AdaptableLayout } from '@kiwicom/mobile-shared';
 
-import ManageMyBookingPackage from './src/appRegistry/ManageMyBookingPackage';
+import NavigationStack from './src/navigation/NavigationStack';
+
+type Props = {|
+  currency: string,
+  accessToken: string,
+|};
+
+class ManageMyBookingPackage extends React.Component<Props> {
+  render = () => (
+    <AdaptableLayout.Provider>
+      <AuthContext.Provider accessToken={this.props.accessToken}>
+        <NavigationStack />
+      </AuthContext.Provider>
+    </AdaptableLayout.Provider>
+  );
+}
 
 AppRegistry.registerComponent('ManageMyBooking', () => ManageMyBookingPackage);
