+++ app/MMB/src/navigation/ListScreen.js
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { HeaderTitle } from '@kiwicom/mobile-navigation';
 import { Translation } from '@kiwicom/mobile-localization';
 
-import Flights from '../list/Flights';
+import Flights from '../scenes/list/Flights';
 
 export default function ListScreen() {
   return <Flights />;
