+++ app/MMB/src/scenes/list/cityImage/CityImage.js
