+++ app/MMB/src/scenes/list/cityImage/ImageBadges.js
