+++ app/MMB/src/scenes/list/cityImage/TravelTypeIcon.js
